feat: Add --update and --help flags
Add an `--update` flag to the `hibernate.sh` script to allow recreation of the swapfile if the system's RAM has changed. This is useful when upgrading the system's memory and needing to adjust the swapfile size accordingly.

Also, add a `--help` flag to the script to display usage information and document the new flags in the `README.md`.